requirements.txt
langchain
langchain-ollama
langchain-chroma
langchain-community
chromadb
pypdf
langchain-text-splitters
python-dotenv